Python API for IRC concatenation
